split the container sniff into its own module
The header walks share nothing with the thumbnailer but the three
content type names they answer with, and no behavior changes here.
